Add rating selection to create gbedu screen

+++ app/src/main/java/com/sanmidev/confamgbedu/ui/gdebuList/GdebuListScreen.kt
@@ -57,7 +57,7 @@ fun GbeduRow(gbedu: Gbedu, onNavigateAction: () -> Unit) {
 
 @Composable
 fun RatingsComponent(rating: Rating) {
-    repeat(rating.value) {
+    repeat(rating.value.toInt()) {
         Image(
             painter = painterResource(R.drawable.ic_ratings_24),
             contentDescription = "Ratings",
